Translating the code in English
Partially translating the code in English and changing the .txt files in .csv files. Creating new method in LoginPage : getTransactions() in order to read data from Transactions.csv file. Bug appeared while implementing this (white spaces at the beginning of the first string) + fixing it by  declaring a substring from the first index to the end.